ARM: at91: sama5d3xek: add cmp board defconfig
The sama5d3x_cmp board which is equipped with the Low-power DDR2 SDRAM,
PMIC ACT8865, and the power rails to measure the power consumption.
NOTE: as those changes are done in at91bootstrap, u-boot need no change.

As there is more power consumption for USB Host clock and USB Device, for
power measurement we disabled USB host and device.

Signed-off-by: Josh Wu <josh.wu@atmel.com>
Acked-by: Bo Shen <voice.shen@atmel.com>