sama5d4: remove default spi flash cs, and bus definition
As if CONFIG_ENV_SPI_BUS, CONFIG_ENV_SPI_CS is not defined, by default
the value is 0.
So we don't need to define it explicitly. If we want to change to
another bus or cs, just need to redefine the CONFIG_ENV_SPI_BUS or
CONFIG_ENV_SPI_CS.

Signed-off-by: Josh Wu <josh.wu@atmel.com>